Refactored into separate files to separate functions
